pyproject.toml
[tool.poetry]
name = "fastapi-finance"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
passlib = "^1.7.4"
python-jose = "^3.3.0"
sqlalchemy = "^2.0.21"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
alembic = "^1.12.0"
pydantic-settings = "^2.0.3"
asyncpg = "^0.28.0"
currencyconverter = "^0.17.11"
yt-dlp = "^2023.10.7"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
sqladmin = "^0.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
